pyproject.toml
[project]
name = "avivo-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.26.0",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "huggingface-hub>=1.7.2",
    "ipykernel>=7.2.0",
    "llama-cpp-python>=0.3.16",
    "llama-index>=0.14.18",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "sqlite-vec>=0.1.7",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
